Move default option implementations to the Core project and adjust namespaces
